Cast port

+++ config/filesystems.php
@@ -63,7 +63,7 @@ return [
             'username' => env('SFTP_USERNAME'),
             'password' => env('SFTP_PASSWORD'),
 
-            'port' => env('SFTP_PORT', 22),
+            'port' => (int) env('SFTP_PORT', 22),
             'root' => env('SFTP_ROOT'),
             'timeout' => 30,
         ],
